Refactor: move from_substrait_extended_expr to mod.rs
